Rename to fit "best practices"
